Save FCSR to pt_regs upon kernel entrance thru handle_exception
This needs to be done because entrance to the kernel through an
exception does NOT preserve ALL FPU state! Only the pieces of HW that
are in pt_regs gets preserved upon entrance to the kernel. Further,
the kernel disables FP support fairly early into the
exception-handling process, which prevents later parts from using ANY
FP-based code, including reading FP CSRs, like FCSR, which makes
determining what FPE_FLTXXX number to send as part of the siginfo.

We must only read & save fcsr if core's FP support is not currently in
the off or initial states.

There is a snafu we hit where if we ALWAYS save fcsr to pt_regs, then
the RCU subsystem does not come up, presumably because the kernel died
before anything useful could get printed out.

```
... Everything is normal to here, then ...
[    0.000712] pid_max: default: 32768 minimum: 301
[    0.000933] LSM: initializing lsm=capability,integrity
[    0.001153] Mount-cache hash table entries: 8192 (order: 4, 65536 bytes, linear)
[    0.001305] Mountpoint-cache hash table entries: 8192 (order: 4, 65536 bytes, linear)
[    0.003839] RCU Tasks Trace: Setting shift to 0 and lim to 1 rcu_task_cb_adjust=1.
[    0.003979] riscv: ELF compat mode unsupported
[    0.003991] ASID allocator disabled (0 bits)
[    0.004205] rcu: Hierarchical SRCU implementation.
[    0.004221] rcu:     Max phase no-delay instances is 1000.
... Linux just hangs here forever ...
```

I figured this could happen under a couple of conditions:
  1. FP support was not previously turned on, and since trying to use
     any FP instructions, registers, and CSRs when FP support is
     turned off causes an illegal instruction exception, perhaps
     something is blowing up.
  2. The core is in some kind of deadlock.

Honestly, the real underlying reason for this issue eludes me, so I
just prevent the core from saving those registers before a later part
of the core has turned on FP support.